HUGE change to src inclusion.
Since the hmmer code needs to be compiled on each machine indiviually to
make sure that the libs are correct, I have moved to a new build system.
Now the Makefile.PL will try and compile the hmmer code base before
creating the perl module Makefile. This is not the most ideal solution,
but it is better than forcing someone to download and install the
correct version of hmmer from subversion. It probably wont work on
windows, but I don't think hmmer will anyway, so meh.