Added two empty .pm modules to make Inline happy.
version 0.56 of the Inline module tried to compile all things in the lib
directory, which caused it to fail, but if we add a .pm file for each of the
files that doesn't have a .pm extension, the install works. This is a
nasty hack that will not be necessary with future versions of Inline.